+++ resources/release-notes.md
@@ -1,4 +1,3 @@
-# Software
+# App
 
-- Fix I-ALiRT `isPlottable` erroring on empty instance
-- Fix undefined `this` in spectrogram plots `mag.*.view.Spectrogram`
+- Add note for unsupported instrument modes in PSD view